Student Registration Project Files/requirements.txt
opencv_python==3.3.1.11
dlib==19.8.1
numpy==1.11.0
cognitive_face==1.3.1
easygui==0.98.1
face_recognition==1.0.0
openpyxl==2.4.9
Pillow==5.1.0
